Support code blocks used in lists, etc.
When parsing the markdown AST, the implementation had assumed that all
`code` nodes in the AST were immediate children of the root node
(`markdownAST`) causing gatsby builds to fail if the markdown source
violated this assumption.

For example, when building markdown like this:

``````
1. First install this thing.
2. Then do this:

   ```js
   const thing = require('installed-thing')
   ```

3. Then start the app.
``````

The gatsby build would fail with:

```
TypeError: Cannot read property 'replace' of undefined
```

This commit fixes this problem by ensuring that the `visit` handler
always uses the actual parent (provided by `unist-util-visit` as the
third argument to the `visit` function) instead of using the root
(`markdownAST`).